feat: wrap logic into mixins
