+++ bugster/libs/mixins.py
@@ -1,0 +1,178 @@
+import os
+import time
+from collections import defaultdict
+
+import yaml
+from rich.console import Console
+from rich.status import Status
+from rich.text import Text
+
+from bugster.constants import TESTS_DIR
+from bugster.libs.utils.diff_parser import parse_git_diff
+from bugster.libs.utils.enums import GitCommand
+from bugster.libs.utils.files import get_specs_paths
+from bugster.libs.utils.git import run_git_command
+from bugster.libs.utils.nextjs.pages_finder import (
+    find_pages_that_use_file,
+    is_nextjs_page,
+)
+
+console = Console()
+
+
+def get_affected_pages(diff_files: list[str], import_tree: dict):
+    """Get the affected pages."""
+    affected_pages = set()
+
+    for file_path in diff_files:
+        if is_nextjs_page(file_path=file_path):
+            affected_pages.add(file_path)
+        else:
+            pages = find_pages_that_use_file(
+                file_path=file_path, import_tree=import_tree
+            )
+
+            if pages:
+                for page in pages:
+                    affected_pages.add(page)
+
+    return affected_pages
+
+
+def get_specs_pages():
+    """Get the specs pages."""
+    specs_files_paths = get_specs_paths()
+    specs_pages = {}
+
+    for spec_path in specs_files_paths:
+        with open(spec_path, encoding="utf-8") as file:
+            data = yaml.safe_load(file)
+            page_path = data["page_path"]
+            relative_path = os.path.relpath(spec_path, TESTS_DIR)
+            specs_pages[page_path] = {
+                "data": data,
+                "path": relative_path,
+            }
+
+    return specs_pages
+
+
+def get_diff_changes_per_page(import_tree: dict):
+    """Get the diff changes per page."""
+    diff_changes_per_page = defaultdict(list)
+    diff_changes = run_git_command(cmd_key=GitCommand.DIFF_CHANGES)
+    parsed_diff = parse_git_diff(diff_text=diff_changes)
+
+    for diff in parsed_diff.files:
+        old_path = diff.old_path
+
+        if is_nextjs_page(file_path=old_path):
+            diff_changes_per_page[old_path] = parsed_diff.to_llm_format(
+                file_change=diff
+            )
+        else:
+            pages = find_pages_that_use_file(
+                file_path=old_path, import_tree=import_tree
+            )
+
+            if pages:
+                for page in pages:
+                    new_diff = parsed_diff.to_llm_format(file_change=diff)
+                    diff_changes_per_page[page].append(new_diff)
+
+    return diff_changes_per_page
+
+
+class UpdateMixin:
+    """Update mixin."""
+
+    def update(self, *args, **kwargs):
+        """Update existing specs."""
+        modified_files = self.mapped_changes["modified"]
+        console.print(f"✓ Found {len(modified_files)} modified files")
+        affected_pages = get_affected_pages(
+            diff_files=modified_files, import_tree=self.import_tree
+        )
+        diff_changes_per_page = get_diff_changes_per_page(import_tree=self.import_tree)
+        updated_specs = 0
+        suggested_specs = []
+        specs_pages = get_specs_pages()
+
+        for page in affected_pages:
+            if page in specs_pages:
+                spec = specs_pages[page]
+                spec_data = spec["data"]
+                spec_path = spec["path"]
+
+                with Status(
+                    f"[yellow]Updating: {spec_path}[/yellow]", spinner="dots"
+                ) as status:
+                    diff = "\n==========\n".join(diff_changes_per_page[page])
+                    self.test_cases_service.update_spec_by_diff(
+                        spec_data=spec_data, diff_changes=diff, spec_path=spec_path
+                    )
+                    status.update(f"✓ [green]{spec_path}[/green] updated")
+                    time.sleep(3)
+                    updated_specs += 1
+            else:
+                text = Text("✗ Page ")
+                text.append(page, style="red")
+                text.append(" not found in test cases")
+                console.print(text)
+
+        if updated_specs > 0 and len(suggested_specs) > 0:
+            console.print(
+                f"✓ Updated {updated_specs} spec{'' if updated_specs == 1 else 's'}, {len(suggested_specs)} suggestion{'' if len(suggested_specs) == 1 else 's'}"
+            )
+        elif updated_specs > 0:
+            console.print(
+                f"✓ Updated {updated_specs} spec{'' if updated_specs == 1 else 's'}"
+            )
+
+
+class SuggestMixin:
+    """Suggest mixin."""
+
+    def suggest(self, *args, **kwargs):
+        """Suggest new specs."""
+        added_files = self.mapped_changes["new"]
+        console.print(f"✓ Found {len(added_files)} added files")
+        suggested_specs = []
+
+        # TODO: Implement the real logic for this
+        import re
+
+        def camel_to_kebab(text):
+            """Convert camelCase to kebab-case."""
+            return re.sub(r"([a-z0-9])([A-Z])", r"\1-\2", text).lower()
+
+        def suggest_spec_from_page(page_path):
+            """Suggest a spec file name from a page path."""
+            clean_path = re.sub(r"^src/pages/", "", page_path)
+            clean_path = re.sub(r"\.(tsx?|jsx?)$", "", clean_path)
+
+            def replace_dynamic(match):
+                param = match.group(1)
+                return camel_to_kebab(param)
+
+            clean_path = re.sub(r"\[([^\]]+)\]", replace_dynamic, clean_path)
+
+            clean_path = camel_to_kebab(clean_path)
+            return f"{clean_path}.yaml"
+
+        for file in added_files:
+            if is_nextjs_page(file_path=file):
+                suggested_specs.append(suggest_spec_from_page(file))
+
+        if len(suggested_specs) > 0:
+            for spec in suggested_specs:
+                console.print(f"⚠️  Suggested new spec: {spec}")
+
+
+class DeleteMixin:
+    """Delete mixin."""
+
+    def delete(self, *args, **kwargs):
+        """Delete existing specs."""
+        deleted_files = self.mapped_changes["deleted"]
+        console.print(f"✓ Found {len(deleted_files)} deleted files")
