+++ bugster/libs/utils/diff_parser.py
@@ -239,115 +239,3 @@ def parse_hunk(lines: List[str], start_index: int) -> Dict[str, Any]:
     }
 
     return {"hunk": hunk_data, "next_index": i}
-
-
-def parse_git_status(status_output):
-    """Parse git status --porcelain output into categorized file lists.
-
-    Git status codes (first char = staged, second char = unstaged):
-    Staged codes (X_):
-    - M = Modified
-    - A = Added (new file)
-    - D = Deleted
-    - R = Renamed
-    - C = Copied
-    - U = Updated but unmerged
-
-    Unstaged codes (_Y):
-    - M = Modified
-    - D = Deleted
-    - R = Type changed (file/symlink/submodule)
-    - U = Updated but unmerged
-    - ? = Untracked
-    - ! = Ignored
-
-    Special cases:
-    - ?? = Untracked file
-    - !! = Ignored file
-
-    :param status_output: Raw output from 'git status --porcelain'.
-    :return: Dictionary with 'modified', 'deleted', and 'new' file lists.
-    """
-    result = {"modified": [], "deleted": [], "new": []}
-
-    # Split by newlines and filter out empty lines
-    lines = [line for line in status_output.strip().split("\n") if line.strip()]
-
-    for line in lines:
-        if len(line) < 3:  # Skip malformed lines
-            continue
-
-        status_code = line[:2]  # First two characters are the status
-        filename = line[3:]  # Rest is the filename (skip the space)
-
-        # Handle special two-character codes first
-        if status_code == "??":
-            # Untracked file - treat as new
-            result["new"].append(filename)
-            continue
-        elif status_code == "!!":
-            # Ignored file - skip (not typically needed for most use cases)
-            continue
-
-        # Get individual status characters
-        staged = status_code[0]
-        unstaged = status_code[1]
-
-        # Determine file category based on status codes
-        # Priority: deleted > new > modified (since a file can have multiple states)
-
-        is_deleted = False
-        is_new = False
-        is_modified = False
-
-        # Check for deletions (highest priority)
-        if staged == "D" or unstaged == "D":
-            is_deleted = True
-
-        # Check for new files
-        elif staged == "A":  # Added/staged new file
-            is_new = True
-        elif staged == "C":  # Copied file (treat as new)
-            is_new = True
-
-        # Check for modifications
-        elif staged == "M" or unstaged == "M":
-            is_modified = True
-        elif staged == "R":  # Renamed (treat as modified)
-            is_modified = True
-        elif unstaged == "R":  # Type changed (treat as modified)
-            is_modified = True
-        elif staged == "U" or unstaged == "U":  # Unmerged (treat as modified)
-            is_modified = True
-
-        # Handle edge cases where both staged and unstaged have status
-        # For example: "AM" = added then modified, "AD" = added then deleted
-        if staged != " " and unstaged != " ":
-            if staged == "A" and unstaged == "D":
-                # Added then deleted - treat as deleted
-                is_deleted = True
-                is_new = False
-            elif staged == "A" and unstaged == "M":
-                # Added then modified - treat as new (since it's still a new file overall)
-                is_new = True
-                is_modified = False
-            elif staged == "M" and unstaged == "D":
-                # Modified then deleted - treat as deleted
-                is_deleted = True
-                is_modified = False
-            elif staged == "D" and unstaged == "M":
-                # This shouldn't normally happen, but treat as modified
-                is_modified = True
-                is_deleted = False
-
-        # Categorize the file
-        if is_deleted:
-            result["deleted"].append(filename)
-        elif is_new:
-            result["new"].append(filename)
-        elif is_modified:
-            result["modified"].append(filename)
-        # If none of the above, it might be an unhandled status code
-        # For robustness, you could add an 'unknown' category or default to 'modified'
-
-    return result
