+++ bugster/libs/utils/enums.py
@@ -5,3 +5,11 @@ class BugsterApiPath(str, Enum):
     """Bugster API paths."""
 
     TEST_CASES = "/api/v1/test-cases"
+
+
+class GitCommand(str, Enum):
+    """Git commands."""
+
+    DIFF_STATUS = "diff_status"
+    DIFF_FILES = "diff_files"
+    DIFF_CHANGES = "diff_changes"
